formula lucro moedas

+++ Bitocin/Content/API/CoinWarzAPI.cs
@@ -20,14 +20,14 @@ namespace Bitocin.Content.API
             public string CoinName { get; set; }
             public string CoinTag { get; set; }
             public string Algorithm { get; set; }
-            public float Difficulty { get; set; }
-            public float BlockReward { get; set; }
+            public Decimal Difficulty { get; set; }
+            public Decimal BlockReward { get; set; }
             public int BlockCount { get; set; }
-            public float ProfitRatio { get; set; }
-            public float AvgProfitRatio { get; set; }
+            public Decimal ProfitRatio { get; set; }
+            public Decimal AvgProfitRatio { get; set; }
             public string Exchange { get; set; }
-            public float ExchangeRate { get; set; }
-            public float ExchangeVolume { get; set; }
+            public Decimal ExchangeRate { get; set; }
+            public Decimal ExchangeVolume { get; set; }
             public bool IsBlockExplorerOnline { get; set; }
             public bool IsExchangeOnline { get; set; }
             public string Message { get; set; }
